moves video module to libraries
