+++ libraries/videoModule/constants/events.js
