+++ libraries/videoModule/constants/vendorCodes.js
