+++ libraries/videoModule/coreVideo.js
