+++ test/spec/modules/videoModule/shared/parentModule_spec.js
@@ -1,4 +1,4 @@
-import { SubmoduleBuilder, ParentModule } from 'modules/videoModule/shared/parentModule.js';
+import { SubmoduleBuilder, ParentModule } from 'libraries/videoModule/shared/parentModule.js';
 import { expect } from 'chai';
 
 describe('Parent Module', function() {
